Allow users to trust external content from certain senders #147
They are also given an option to revoke the trust. I am sort of concerned about the possibility of too many banners stacking if this one triggers alongside, eg a phishing warning banner. Maybe they could be combined or flattened somehow? Otherwise this could easily be converted to some buttons. This commit also includes some cleanup and modification of Banner and LanguageViewModel to (hopefully) increase uniformity and type safety